# docs/api/openapi.locked.yaml
openapi: 3.1.0
info:
  title: FaultMaven API
  description: AI-powered troubleshooting assistant for Engineers, SREs, and DevOps
    professionals
  version: 1.0.0
paths:
...
  /api/v1/knowledge/documents:
    post:
      tags:
      - knowledge_base
      - knowledge_base
      summary: Upload Document
      description: "Upload a document to the knowledge base\n\nArgs:\n    file: Document\
        \ file to upload\n    title: Document title\n    document_type: Type of document\n\
        \    tags: Comma-separated tags\n    source_url: Source URL if applicable\n\
        \nReturns:\n    Upload job information"
      operationId: upload_document_api_v1_knowledge_documents_post
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_document_api_v1_knowledge_documents_post'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Upload Document Api V1 Knowledge Documents Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - knowledge_base
      - knowledge_base
      summary: List Documents
      description: "List knowledge base documents with optional filtering\n\nArgs:\n\
        \    document_type: Filter by document type\n    tags: Filter by tags (comma-separated)\n\
        \    limit: Maximum number of documents to return\n    offset: Number of documents\
        \ to skip\n\nReturns:\n    List of documents"
      operationId: list_documents_api_v1_knowledge_documents_get
      parameters:
      - name: document_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Document Type
      - name: tags
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Tags
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response List Documents Api V1 Knowledge Documents Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/knowledge/documents/{document_id}:
    get:
      tags:
      - knowledge_base
      - knowledge_base
      summary: Get Document
      description: "Get a specific knowledge base document\n\nArgs:\n    document_id:\
        \ Document identifier\n\nReturns:\n    Document details"
      operationId: get_document_api_v1_knowledge_documents__document_id__get
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: string
          title: Document Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeBaseDocument'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - knowledge_base
      - knowledge_base
      summary: Delete Document
      description: "Delete a knowledge base document\n\nArgs:\n    document_id: Document\
        \ identifier\n\nReturns:\n    Deletion confirmation"
      operationId: delete_document_api_v1_knowledge_documents__document_id__delete
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: string
          title: Document Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - knowledge_base
      - knowledge_base
      summary: Update Document
      description: Update document metadata and content.
      operationId: update_document_api_v1_knowledge_documents__document_id__put
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: string
          title: Document Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: Update Data
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Update Document Api V1 Knowledge Documents  Document
                  Id  Put
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/knowledge/jobs/{job_id}:
    get:
      tags:
      - knowledge_base
      - knowledge_base
      summary: Get Job Status
      description: "Get the status of a knowledge base ingestion job\n\nArgs:\n  \
        \  job_id: Job identifier\n\nReturns:\n    Job status information"
      operationId: get_job_status_api_v1_knowledge_jobs__job_id__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Get Job Status Api V1 Knowledge Jobs  Job Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/knowledge/search:
    post:
      tags:
      - knowledge_base
      - knowledge_base
      summary: Search Documents
      description: "Search knowledge base documents\n\nArgs:\n    request: Search\
        \ request with query and filters\n\nReturns:\n    Search results"
      operationId: search_documents_api_v1_knowledge_search_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Search Documents Api V1 Knowledge Search Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/knowledge/documents/bulk-update:
    post:
      tags:
      - knowledge_base
      - knowledge_base
      summary: Bulk Update Documents
      description: Bulk update document metadata.
      operationId: bulk_update_documents_api_v1_knowledge_documents_bulk_update_post
      requestBody:
        content:
          application/json:
            schema:
              type: object
              title: Request
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Bulk Update Documents Api V1 Knowledge Documents Bulk
                  Update Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/knowledge/documents/bulk-delete:
    post:
      tags:
      - knowledge_base
      - knowledge_base
      summary: Bulk Delete Documents
      description: Bulk delete documents.
      operationId: bulk_delete_documents_api_v1_knowledge_documents_bulk_delete_post
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties:
                items:
                  type: string
                type: array
              type: object
              title: Request
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Bulk Delete Documents Api V1 Knowledge Documents Bulk
                  Delete Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/kb/documents:
    post:
      tags:
      - knowledge_base
      - knowledge_base
      summary: Upload Document Kb
      description: Upload a document to the knowledge base (kb prefix)
      operationId: upload_document_kb_api_v1_kb_documents_post
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_document_kb_api_v1_kb_documents_post'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Upload Document Kb Api V1 Kb Documents Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - knowledge_base
      - knowledge_base
      summary: List Documents Kb
      description: List knowledge base documents (kb prefix)
      operationId: list_documents_kb_api_v1_kb_documents_get
      parameters:
      - name: document_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Document Type
      - name: tags
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Tags
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response List Documents Kb Api V1 Kb Documents Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/kb/jobs/{job_id}:
    get:
      tags:
      - knowledge_base
      - knowledge_base
      summary: Get Job Status Kb
      description: Get job status (kb prefix)
      operationId: get_job_status_kb_api_v1_kb_jobs__job_id__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Get Job Status Kb Api V1 Kb Jobs  Job Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/kb/search:
    post:
      tags:
      - knowledge_base
      - knowledge_base
      summary: Search Documents Kb
      description: Search documents (kb prefix)
      operationId: search_documents_kb_api_v1_kb_search_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Search Documents Kb Api V1 Kb Search Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/kb/documents/bulk-update:
    post:
      tags:
      - knowledge_base
      - knowledge_base
      summary: Bulk Update Documents Kb
      description: Bulk update documents (kb prefix)
      operationId: bulk_update_documents_kb_api_v1_kb_documents_bulk_update_post
      requestBody:
        content:
          application/json:
            schema:
              type: object
              title: Request
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Bulk Update Documents Kb Api V1 Kb Documents Bulk
                  Update Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_upload_document_api_v1_knowledge_documents_post:
      properties:
        file:
          type: string
          format: binary
          title: File
        title:
          type: string
          title: Title
        document_type:
          type: string
          title: Document Type
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
        tags:
          anyOf:
          - type: string
          - type: 'null'
          title: Tags
        source_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Url
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      type: object
      required:
      - file
      - title
      - document_type
      title: Body_upload_document_api_v1_knowledge_documents_post
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...